+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Support the latest version of `package:analyzer`.
 * Fix hyperlinks to overriden methods (#1994).
 * Option in dartdoc_options.yaml to exclude version in footer info (#1982).
+* Sanitize HTML output of user-provided markdown content.
 
 ## 0.28.4
 * **Breaking change** Change the default for `allow-tools` command line flag to false.